refactor: pull out search types into new file
