+++ site/search/HeaderSearch.tsx
@@ -2,7 +2,8 @@ import React from "react"
 import { observable, action, runInAction } from "mobx"
 import { observer } from "mobx-react"
 import { SearchResults } from "./SearchResults.js"
-import { SiteSearchResults, siteSearch } from "./searchClient.js"
+import { siteSearch } from "./searchClient.js"
+import { SiteSearchResults } from "./searchTypes.js"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome/index.js"
 import { faSearch } from "@fortawesome/free-solid-svg-icons/faSearch"
