+++ site/search/searchClient.ts
@@ -1,9 +1,15 @@
 import algoliasearch, { SearchClient } from "algoliasearch/lite"
-import { countries, Country } from "@ourworldindata/utils"
+import { countries } from "@ourworldindata/utils"
 import {
     ALGOLIA_ID,
     ALGOLIA_SEARCH_KEY,
 } from "../../settings/clientSettings.js"
+import {
+    ArticleHit,
+    CountryHit,
+    SiteSearchResults,
+    ChartHit,
+} from "./searchTypes.js"
 
 let algolia: SearchClient | undefined
 const getClient = () => {
@@ -11,62 +17,7 @@ const getClient = () => {
     return algolia
 }
 
-type PageHit = ArticleHit | CountryHit
-
-export interface CountryHit {
-    objectID: string
-    type: "country"
-    slug: string
-    title: string
-    code: string
-    content: string
-    _highlightResult: any
-    _snippetResult: {
-        content: {
-            value: string
-        }
-    }
-}
-
-export interface ArticleHit {
-    objectID: string
-    postId: number
-    slug: string
-    title: string
-    type: "post" | "page" | "entry" | "explainer" | "fact"
-    content: string
-    _snippetResult: {
-        content: {
-            value: string
-        }
-    }
-}
-
-export interface ChartHit {
-    chartId: number
-    slug: string
-    title: string
-    subtitle: string
-    variantName: string
-    _snippetResult?: {
-        subtitle: {
-            value: string
-        }
-    }
-    _highlightResult?: {
-        availableEntities: {
-            value: string
-            matchLevel: "none" | "full"
-            matchedWords: string[]
-        }[]
-    }
-}
-
-export interface SiteSearchResults {
-    pages: PageHit[]
-    charts: ChartHit[]
-    countries: Country[]
-}
+export type PageHit = ArticleHit | CountryHit
 
 export const siteSearch = async (query: string): Promise<SiteSearchResults> => {
     // Some special ad hoc handling of country names for chart query
